+++ examples/io/display.py
@@ -19,3 +19,12 @@ if __name__ == "__main__":
         print(f"Frame {frame_id}: {frame.shape}")
 
     display.stop()
+
+    # OR
+    # use with context manager
+
+    video = IterableVideo("video.mp4")
+    with Display("example") as display:
+        for frame_id, frame in video:
+            display(frame)
+            print(f"Frame {frame_id}: {frame.shape}")