+++ src/cv2ext/bboxes/filters/__init__.py
@@ -2,6 +2,7 @@
 #
 # MIT License
 """Submodule implementing filters for bounding boxes."""
+
 from __future__ import annotations
 
 from ._kalman import KalmanFilter